Add three macros, quantifying DAB in a subarea
A series of three subsequent macros, using IHC images, stained with DAB to show protein expression as well as segmentation masks of subareas (i.e. tissue component), obtained with ilastik or other deep learning models. The aim is to quantify the mean intensity per subarea and then, the ratios between the mean intensities of different subareas as a semiquantitative measure of distribution.